admin page

+++ src/repositories/UserRepo.php
@@ -4,15 +4,17 @@ require_once 'src/repositories/Repo.php';
 require_once 'src/models/User.php';
 
 class UserRepo extends Repo {
-    public function createUser($email, $username, $password, $name, $surname, $room_number) {
-        $sql = "INSERT INTO users (email, login, password, name, surname, room_number) VALUES (:email, :username, :password, :name, :surname, :room_number)";
+    public function createUser($email, $username, $password, $name, $surname, $room_number, $user_role = "user", $user_blocked = false) {
+        $sql = "INSERT INTO users (email, login, password, name, surname, room_number, user_role, user_blocked) VALUES (:email, :username, :password, :name, :surname, :room_number: user_role, user_blocked)";
         $params = array(
             ':email' => $email,
             ':username' => $username,
             ':password' => password_hash($password, PASSWORD_DEFAULT),
             ':name' => $name,
             ':surname' => $surname,
-            ':room_number' => $room_number
+            ':room_number' => $room_number,
+            ':user_role' => $user_role,
+            ':user_blocked' => $user_blocked,
         );
 
         $this->databaseController->execute($sql, $params);
@@ -32,14 +34,28 @@ class UserRepo extends Repo {
     }
 
     public function deleteUser($userId) {
-        $sql = "DELETE FROM users WHERE userid = :userid";
+        $sql = "DELETE FROM users WHERE userid = :userid AND user_role != 'admin'";
         $params = array(':userid' => $userId);
 
         $this->databaseController->execute($sql, $params);
     }
 
+    public function blockUser($userId) {
+        $sql = "UPDATE users SET user_blocked = true WHERE user_id = :userId";
+        $params = array(':userId' => $userId);
+
+        $this->databaseController->execute($sql, $params);
+    }
+
+    public function unblockUser($userId) {
+        $sql = "UPDATE users SET user_blocked = false WHERE user_id = :userId";
+        $params = array(':userId' => $userId);
+
+        $this->databaseController->execute($sql, $params);
+    }
+
     public function getUser($userId) {
-        $sql = "SELECT * FROM users WHERE user_id = :userId";
+        $sql = "SELECT * FROM users WHERE user_id = :userId AND user_role != 'admin'";
         $params = array(':userId' => $userId);
 
         $stmt = $this->databaseController->execute($sql, $params);
@@ -53,11 +69,37 @@ class UserRepo extends Repo {
                 $user['password'],
                 $user['name'],
                 $user['surname'],
-                $user['room_number']
+                $user['room_number'],
+                $user['user_role'],
+                $user['user_blocked']
             );
         }
 
         return null;
+    }
+
+    public function getUsers() {
+        $sql = "SELECT * FROM users WHERE user_role != 'admin'";
+        $stmt = $this->databaseController->execute($sql, null);
+
+        $users = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        $usersArray = array();
+
+        foreach($users as $user) {
+            $usersArray[] = new User(
+                $user['user_id'],
+                $user['email'],
+                $user['login'],
+                $user['password'],
+                $user['name'],
+                $user['surname'],
+                $user['room_number'],
+                $user['user_role'],
+                $user['user_blocked']
+            );
+        }
+
+        return $usersArray;
     }
 
     public function getUserByLogin($login) {
@@ -75,7 +117,9 @@ class UserRepo extends Repo {
                 $user['password'],
                 $user['name'],
                 $user['surname'],
-                $user['room_number']
+                $user['room_number'],
+                $user['user_role'],
+                $user['user_blocked']
             );
         }
